+++ plugin-code/src/main/java/com/trivago/cluecumber/rendering/pages/pojos/pagecollections/LinkType.java
@@ -1,0 +1,6 @@
+package com.trivago.cluecumber.rendering.pages.pojos.pagecollections;
+
+public enum LinkType {
+    INTERNAL,
+    EXTERNAL
+}