+++ CHANGELOG.md
@@ -7,11 +7,18 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 Back to [Readme](README.md).
 
-## [2.8.1] - UNRELEASED
+## [2.8.1] - 2022-09-23
+
+### Added
+* Ability to add custom external links to the navigation bar via `customNavigationLinks` property
+
+### Fixed
+* #281 - enable Cluecumber to be invoked without a POM project
 
 ### Changed
 * Dependency updates
 * Design adjustments
+* Reworked navigation for future improvements
 
 ## [2.8.0] - 2022-01-18
 
@@ -634,6 +641,7 @@ steps with status `pending` or `undefined` (default value is `false`) (#74)
 
 Initial project version on GitHub and Maven Central.
 
+[2.8.1]: https://github.com/trivago/cluecumber-report-plugin/tree/2.8.1
 [2.8.0]: https://github.com/trivago/cluecumber-report-plugin/tree/2.8.0
 [2.7.1]: https://github.com/trivago/cluecumber-report-plugin/tree/2.7.1
 [2.7.0]: https://github.com/trivago/cluecumber-report-plugin/tree/2.7.0